Server/requirements.txt
flask
flask-cors
requests
geopandas
osm2geojson>=0.2.0


# per installare, usa "pip install -r requirements.txt"
